Try to parse blocks as a sequence of 11-bytes records
The tail of the blocks is a crc.
In a record:
The first byte seems to contain 2 things:
 - the 2 MSB bits are some flags
 - with the flags at 0 the rest of the byte is an increasing sequence
(multiple records in a row can have the same counter value)
For such flag-less records, the current theory is a 2 bytes type
identifier followed by 8 bytes payload.
The records with flags could be some control data, 0x40 appears at the
start of each block and whenever the sequence counter is overflowing.
Some record both bits set (0xc0) also sometimes appear.